deprecated_jobs/dwh/job_language/requirements.txt
pandas==1.4.3
psycopg2==2.9.3
python-dotenv==0.21.1
SQLAlchemy==1.4.40
slack_sdk